Piping Tools version 1.9 (26 June 2013)
Updated the followings files:
	~/php/setup.php
	~/php/setting.php
	~/php/dwc_page.php
	~/php/scheduler.php
	~/php/scheduler_pmonit.php

This version have the followings changes:
* Cleaned up setup.php by removing most of the hard coded "piping_devel".

  Affected files:
	. setup.php
	. setting.php
	. dwc_page.php

* Updated scheduler.php and scheduler_pmonit.php to avoid warning messages
  when directories are being removed.

  Affected files:
	. scheduler.php
	. scheduler_pmonit.php

+++ php/scheduler.php
@@ -240,7 +240,7 @@ $templateTxt =
 `provider` varchar(24),
 `gsd_comments` varchar(500) COLLATE utf8_general_ci DEFAULT NULL,
 `gsd_comments_predefined` enum('comment1','comment2','comment3') COLLATE utf8_general_ci DEFAULT NULL,
-`gsd_short_name` varchar(255) COLLATE utf8_general_ci DEFAULT NULL,
+`gsd_short_name` varchar(500) COLLATE utf8_general_ci DEFAULT NULL,
 `gsd_status` enum('Approved','Rejected') DEFAULT NULL,
 `history_status` int(10) DEFAULT NULL,
 `history_comments` varchar(255) COLLATE utf8_general_ci DEFAULT NULL,
@@ -396,6 +396,9 @@ function rm_dir($dir)
 	foreach (new  RecursiveIteratorIterator( $iter,
 						RecursiveIteratorIterator::CHILD_FIRST) as $f)
 	{
+		$filename = $f->getFilename();
+		if ($filename === '.' || $filename === '..')
+			continue;
 		if ($f->isDir())
 			rmdir($f->getPathname());
 		else
